# api: add field which caches content of LabelSelector string representation of EndpointSelector