--- a/pkg/policy/api/ingress.go
+++ b/pkg/policy/api/ingress.go
@@ -145,6 +145,9 @@ func (i *IngressRule) GetSourceEndpointSelectorsWithRequirements(requirements []
 			sel := *i.FromEndpoints[idx].DeepCopy()
 			sel.MatchExpressions = append(sel.MatchExpressions, requirements...)
 			sel.SyncRequirementsWithLabelSelector()
+			// Even though this string is deep copied, we need to override it
+			// because we are updating the contents of the MatchExpressions.
+			sel.cachedLabelSelectorString = sel.LabelSelector.String()
 			res = append(res, sel)
 		}
 	} else {