Performance analysis of policy generatinon revealed that there was a large
amount of time being spent calling \`String()\` on the LabelSelector
within the EndpointSelector, as the string representation of the LabelSelector
within an EndpointSelector is used to lookup the set of identities matched
by said EndpointSelector in the SelectorCache. To reduce the performance
impact of creation of strings upon policy computation, store the string
representation of the LabelSelector as a non-exported field in the
EndpointSelector.

This field is populated whenever we create an EndpointSelector in Cilium.
This will consume more memory overall in the cilium-agent, but the performance
improvements gained by caching this string representation are considerable:

Before:

```
PASS: resolve_test.go:166: PolicyTestSuite.BenchmarkRegeneratePolicyRules	     200	   9003593 ns/op
```

```
ROUTINE ======================== github.com/cilium/cilium/pkg/policy.(*SelectorCache).AddIdentitySelector in /Users/ianvernon/go/src/github.com/cilium/cilium/pkg/policy/selectorcache.go
      40ms      1.49s (flat, cum) 49.34% of Total
         .          .    592:func (sc *SelectorCache) AddIdentitySelector(user CachedSelectionUser, selector api.EndpointSelector) (cachedSelector CachedSelector, added bool) {
         .          .    593:	// The key returned here may be different for equivalent
         .          .    594:	// labelselectors, if the selector's requirements are stored
         .          .    595:	// in different orders. When this happens we'll be tracking
         .          .    596:	// essentially two copies of the same selector.
         .      600ms    597:	key := selector.LabelSelector.String()
         .       10ms    598:	sc.mutex.Lock()

```

After:

```
PASS: resolve_test.go:166: PolicyTestSuite.BenchmarkRegeneratePolicyRules	     500	   6230246 ns/op
```

```
      70ms      1.18s (flat, cum) 29.80% of Total
         .          .    593:	// The key returned here may be different for equivalent
         .          .    594:	// labelselectors, if the selector's requirements are stored
         .          .    595:	// in different orders. When this happens we'll be tracking
         .          .    596:	// essentially two copies of the same selector.
         .          .    597:	key := selector.CachedString()
```

This is approximately a 31 % performance increase in the policy computation
benchmark.

Signed-off by: Ian Vernon <ian@cilium.io>

Fixes: #8093 

<!-- Reviewable:start -->
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cilium/cilium/8365)
<!-- Reviewable:end -->